refactor: standardize Lua module structure
This follows the [Neovim Lua Guide](https://neovim.io/doc/user/lua-guide.html#lua-guide-modules) on how
to structure a module. This enables the project to be used easily with
plugin managers like lazy.nvim
