+++ lua/neotest-kotlin/src/treesitter/spec-query.lua
